feat: FruitStore->FruitRepository 리네이밍, readStock 메소드 구현
